requirements.txt
requests~=2.32.3
urllib3~=2.2
pandas~=2.2
tqdm~=4.66
python-dotenv~=1.0.1
